add a namespace to frontend urls
