--- app/controllers/authentication_controller.rb
@@ -1,15 +1,0 @@
-class AuthenticationController < ApplicationController
-  skip_before_action :verify_authenticity_token 
-  skip_before_action :authenticate_request
- 
-  def authenticate
-    
-    command = AuthenticateUser.call(params[:phone], params[:password])
- 
-    if command.success?
-      render json: { auth_token: command.result }
-    else
-      render json: { error: command.errors }, status: :unauthorized
-    end
-  end
- end
