+++ app/controllers/pages_controller.rb
@@ -1,6 +1,7 @@
+
 class PagesController < ApplicationController
   skip_before_action :authenticate_request
 
   def index
   end
-end
+end
